Machine-Learning-Code-5487/Code_Assignment2/requirements.txt
pandas
numpy
jupyter
scipy
matplotlib
seaborn
pillow
scikit-learn